post management

+++ users/models.py
@@ -14,16 +14,20 @@ class Profile(models.Model):
     def __str__(self):
         return f"{self.user.username} Profile"
     
-    def save(self):
-        old_image = Profile.objects.get(pk=self.pk).image
-        if self.image and self.image != old_image:
-            self.image.name = profile_image_path(self, self.image.name)
+    def save(self, *args, **kwargs):
+        if self.pk:
+            try:
+                old_image = Profile.objects.get(pk=self.pk).image
+            except Profile.DoesNotExist:
+                old_image = None
+            if self.image and self.image != old_image:
+                self.image.name = profile_image_path(self, self.image.name)
+                super().save()
+                img = Image.open(self.image.path)
+                if(img.height > 200 or img.width > 200):
+                    output_size = (200, 200)
+                    img.thumbnail(output_size)
+                    img.save(self.image.path)
+                if old_image != 'default.jpg' and os.path.isfile(old_image.path):
+                        os.remove(old_image.path)
         super().save()
-        img = Image.open(self.image.path)
-        if(img.height > 200 or img.width > 200):
-            output_size = (200, 200)
-            img.thumbnail(output_size)
-            img.save(self.image.path)
-        if old_image and old_image != self.image:
-            if os.path.isfile(old_image.path):
-                os.remove(old_image.path)
